Reorder functions of the pipeline queryText
